+++ src/test/kotlin/me/bossm0n5t3r/leetcode/palindromeLinkedList/PalindromeLinkedListTest.kt
@@ -1,12 +1,12 @@
-package palindromeLinkedList
+package me.bossm0n5t3r.leetcode.palindromeLinkedList
 
+import me.bossm0n5t3r.leetcode.utils.ListNode
+import me.bossm0n5t3r.leetcode.utils.ListNodeUtil
 import org.junit.jupiter.api.Test
-import utils.ListNode
-import utils.ListNodeUtil
 import kotlin.test.assertEquals
 
 class PalindromeLinkedListTest {
-    private val palindromeLinkedList = PalindromeLinkedList()
+    private val palindromeLinkedList = PalindromeLinkedList.Solution()
     private val listNodeUtil = ListNodeUtil()
 
     data class PalindromeLinkedListTestData(
